int for linkcount. If there are more than 2631 links, we are in trouble anyway
git-svn-id: https://svn.code.sf.net/p/autowikibrowser/code/AWB@7581 5227b50d-9349-4574-b055-4b0a7525e11c

+++ WikiFunctions/Parse/Parsers.cs
@@ -5769,7 +5769,7 @@ namespace WikiFunctions.Parse
             articleText = TagEmptySection(articleText);
 
             int totalCategories;
-            double linkCount = Tools.LinkCount(commentsStripped);
+            int linkCount = Tools.LinkCount(commentsStripped);
 
             #if DEBUG || UNITTEST
             if (Globals.UnitTestMode)
